Agent Service

+++ src/main/java/com/service/agent/AgentServiceImpl.java
@@ -1,11 +1,11 @@
 package com.service.agent;
 
 import com.dao.AgentDAO;
+import com.utils.Error;
 import com.model.Agent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 
 @Service
@@ -19,8 +19,15 @@ public class AgentServiceImpl implements AgentService {
         return agentDAO.findAll();
     }
 
-    public Agent findById(Long id) {
-        return agentDAO.findOne(id);
+    public Object findById(Long id)
+    {
+        Agent agnt = new Agent();
+        agnt = agentDAO.findOne(id);
+        if(agnt == null)
+        {
+            return (new Error("Null value", "null_value", 400));
+        }
+        return agnt;
     }
 
     public Agent addAgent(Agent agent)
@@ -34,10 +41,14 @@ public class AgentServiceImpl implements AgentService {
         return agentDAO.save(agnt);
     }
 
-    public Agent deleteAgent(Long id)
+    public Object deleteAgent(Long id)
     {
         Agent agnt = new Agent();
         agnt = agentDAO.findOne(id);
+        if(agnt == null)
+        {
+            return (new Error("Null value", "null_value", 400));
+        }
         agentDAO.delete(id);
         return agnt;
     }
